migrate to main dashboard for create
